Update README.md
In our findings we found the failed kickstarter campaign was unsuccessful due to asking for too much money, so Louise should look at what is causing people not to pledge to the unsuccessful plays. Louise can have a play produced in GB if she is willing to change her budget under $4000. Overall, Theater appears to be a great market to branch into.